♻️ Extract createEmbeddedConfigSetup to shared helper to avoid cross-scenario imports
Playwright prohibits importing between test scenario files. Move the shared
setup factory to test/e2e/lib/helpers/embeddedConfigSetup.ts so both
embeddedConfig and embeddedConfigDynamic scenarios can import it cleanly.
